Added new class to generate sizes
