+++ cs/TagsCloudTests/TagCloudCreatorShould.cs
@@ -15,8 +15,11 @@ public class TagCloudCreatorShould
     [Repeat(5)]
     public void Create_ReturnTagCloudEquivalentToConstructor_AfterExecution()
     {
-        var rectangleSizes = Enumerable.Range(_random.Next(2, 10), _random.Next(20, 100)).
-            Select(width => new Size(width, width / 2)).Reverse().ToArray();
+        var sizesGenerator = new RandomSizesGenerator();
+        var rectangleSizes = sizesGenerator
+            .GenerateSize()
+            .Take(_random.Next(10, 200))
+            .ToArray();
         var currCloudLayouter = new CircularCloudLayouter(_defaultCenter);
         var expectedCloudLayouter = new CircularCloudLayouter(_defaultCenter);
         var defaultTagCloud = new TagCloud(currCloudLayouter);
