+++ cs/TagsCloudVisualization/Interfaces/ISizesGenerator.cs
@@ -1,0 +1,8 @@
+﻿using System.Drawing;
+
+namespace TagsCloudVisualization.Interfaces;
+
+public interface ISizesGenerator
+{
+    public IEnumerable<Size> GenerateSize();
+}
